requirements.txt
pandas
pyyaml
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-ollama
langchain-core
pytest